Split models/game.ts into per-concern mixin modules
game.ts had grown to ~516 lines mixing getters, mutations, actions, and
persistence in a single TitanGame class. The Vuex store wires itself up by
reflecting over TitanGame.prototype (Object.getOwnPropertyNames, filtered by
get*/m*/do* prefixes), so a plain multi-file split isn't possible without
either changing the store or losing that reflection.

Splitting the class body into mixin objects (getters.ts, mutations.ts,
actions.ts, persistence.ts) applied onto TitanGame.prototype via
Object.assign preserves the exact prototype shape the store depends on,
while TypeScript declaration merging keeps the class's public interface
type-checked. Verified via a throwaway build comparing
Object.getOwnPropertyNames(TitanGame.prototype) and the store's derived
getters/mutations/actions maps before and after — identical in both cases.

No behavior change; code moved, not rewritten.
